requirements.txt
Django~=4.1.4
djangorestframework~=3.14.0
whitenoise~=6.2.0